rename CRDs from Piraeus* to Linstor*
Piraeus is the name of the open source project/distribution that makes
LINSTOR available in Kubernetes. The underlying technology that the
Piraeus operator manages is still LINSTOR. The name of the custom
resource definition should match with what it represents, in this case
LINSTOR components. Hence the CRD name needs to be changed.