allow the DB connection URL to be overridden in the Helm values
The property in LinstorControllerSet is still called etcdURL because
renaming it would requite an API version change. Almost all the time it
will really be the URL of the etcd service. However, it can also be used
for different DB types supported by LINSTOR and is passed directly
through to LINSTOR as the db.connection_url config parameter.