update golinstor dependency for better toml writing
golinstor v0.25.1 includes toml tags for ControllerConfig.
Previously, the serialized TOML would contain every item with
a default value. However, LINSTOR does not care that these are
placeholders and so tries to load all values as given.

Interestingly, LINSTOR is able to start with these settings, but
for it will start using an in-memory database, as the "in_memory" field
is non-empty.