refactor secure comms between linstor components
Use hardcoded password "linstor" for java keystores. This is not a big
compromise in security, as the passwords would have been available in the
linstor*set resources, which are generally less protected than secrets.

This commit removes all code handling different passwords for the keystore
and simplifies handling of SSL connector configuration in the operator.