linstor: bind-mount /sys/ in satellites for NVMe support
linstor satellites supporting NVMe need access to:
* /sys/kernel/config/nvmet/
* /sys/devices/virtual/nvme-fabrics/ctl/nvme
While the second entry is mounted from sysfs, which is already
present in the container, the first entry is part of the configfs
mount "/sys/kernel/config", which is not mounted by default.

With this change, linstor-satellite will bind-mount /sys from the host.
As long as the host has the required kernel modules loaded for NVMe,
linstor will be able to control the NVMe layer.