requeue reconcile loop at expected wait points
At some points we need to wait for specific resource to come online.
* Pods need to be scheduled
* Services need to come online (and responsive)

This change adds a new TemporaryError wrapper around a standard error.
It can be converted to a reconcile result, that automatically requeues
the reconcile loop after a given duration elapsed. We currently use it in:

* the controller, waiting until the pod is running (_NOT_ ready)
* the controller, waiting until the controller responds
* the nodeset, waiting until the controller is reachable
* the nodeset, waiting until the node is online in the controller

A few helper methods are also added, to deal with mixed temporary and
"normal" errors. A temporary error takes precedence over normal errors,
i.e. if in a collection of errors a temporary one exists, the combined
return value with be (<requeue>, <no-error>)