build: Build in Docker with operator-sdk v0.18.1
Version v0.18 of operator-sdk introduced a new structure that allows
builds to run entirely in Docker without a go build environment on the
host. This is useful for CI, so switch to using it. The new structure
seems to be undocumented, but it uses kubebuilder. These changes are
based on a basic kubebuilder project.