use deployment to configure linstor-controller
By using a deployment, a controller can be re-scheduled in case of a node
failure without human intervention. Since multiple LINSTOR controllers
cannot manage the same set of nodes at the same time, we cap the replica
count at 1 and use the "Recreate" deployment strategy.

In case of a node going suddenly offline, a new pod is scheduled after the
default timeout (300s). The old pod will still be marked as running, but it
won't receive any traffic (the node is offline, and the service endpoint would
not route traffic in any case).

The possibility of multiple pods associated with a controller deployment being
available in the Kubernetes API at the same time needed some rework in the
ControllerSet registration and status update code.