add special resource specification for injector image
While working on #53, it became clear that there is no way to set
the resources for the injector init container. This init container
is special however, as the range of required resources varies between
different injection modes.

The best solution is: let the user decide. This commit adds a new
value to the satelliteSet resource "kernelModuleInjectionResources".
This is directly passed to the init container.

The example resource settings are updated to reflect a known-good
resource request for compiling the kernel module from source.