Upgrade CRDs to v1
Upgrading the CRDs to v1 marks them as stable, which is a requirement
for the v1.0 release of the operator.

The actual types are simple copies of the current v1alpha1 versions.
They are marked as "storageversions" i.e. kubernetes will store all
resources as v1.

The operator itself already works on v1 resources, except for the
legacy controller (which will write v1alpha versions that are directly
converted to v1 by kubernetes).

To ensure that v1alpha1 is compatible with v1, a simple test serializes
v1alpha types and deserializes them as v1.