only start legacy controllers if legacy CRDs present
Previously, if legacy CRDs are not present the legacy contoller cannot start.
This aborts the whole operator runtime. Instead we want to continue running
without the legacy controllers.

This commit adds a manual check for legacy CRDs to each legacy controller.
The extra check is needed because the controller ".Watch()" function does
no sanity check in this regard. The error is generated by the controller
runtime at a later point.