Helm: make creation of controller and satellite set optional
In some cases you want to manually deploy the controller and/or satellites, or
manually add the resources at a later time. This commit adds the option to skip
creation of LinstorController and LinstorSatelliteSet resources (was already
possible for LinstorCSIDriver).

Also adds the option to set the LINSTOR controller URL to any desired value.