fix helm warnings when merging tables
For Helm a "null" merged with a "table" is a warning. Instead of "null"
we should use empty tables.

However, we still want to differentiate between the default value and an
intentional empty value. This is important for the "affinity" settings,
as we use a default value if go encounters as "nil" value. To support this, we
leave out the whole key if the default value is encountered in the Helm
template.

The effect of this in the operator code is:
* A value of "{}" will be seen as "nil". This will trigger setting the default
  affinity.
* A value that only contains a table key but no table value will be passed
  through. The effect is that no affinity will be configured, and we won't
  override it with the default affinity.