Update operator-sdk to v0.19 + LINSTOR client v0.33
CRDs with apiVersion v1beta are deprecated starting with kubernetes v1.19
and will be completely removed in v1.22. The operator-sdk starting with
v0.18 will generate the newer v1 CRDs.

We update operator-sdk to the latest <1.0 version to stay compatible with our
current project layout. A complete switch to the latest operator-sdk is desired
but currently deemed to time consuming.

Additionally, we update the LINSTOR client to v0.33