update default docker images + fix upgrade issues
Update all CSI sidecars and external images to the latest version
supporting Kubernetes 1.17 and up.

Also fix an issue when using Helm to upgrade to the latest release:
Helm would do it's own CRD validation, which would quickly fail when
setting the "additionalEnv" and "additionalProperties" on clusters
with old CRDs. As a workaround, we only include those props in the
template if non-empty. This has the benefit that setting these with
old CRDs deployed (which would silently remove those values) will fail.

Document the upgrade procedure for using the new CRD values

Signed-off-by: Moritz "WanzenBug" Wanzenböck <moritz.wanzenboeck@linbit.com>