Use recommended labels for created resources
Replaces the different labels applied to deployments, daemonsets, etc. with
a set of recommended labels [1]. This makes it easier to track which resources
are created by the operator.

Updating the existing deployments and daemonsets requires some extra work:
the label selector used internally by the workload manager is immutable.
Updates to the pod template labels and label selector gets rejected by the
API. As a workaround, we first delete the resource and recreate it with the
new labels afterwards.

[1]: https://kubernetes.io/docs/concepts/overview/working-with-objects/common-labels/

Signed-off-by: Moritz "WanzenBug" Wanzenböck <moritz.wanzenboeck@linbit.com>