delete auxiliary satellite props if not in labels
In 530c7ae, the operator started applying node labels to the LINSTOR satellite
objects as auxiliary properties. It did not support deleting these labels
once they were removed from the Kubernetes node. This caused an issue
with the CSI Driver, which used the auxiliary properties when advertising
the supported scheduling labels. When it reported a protected label (i.e.
"kubernetes.io/...") that was no longer present, the CSI Driver failed to
start.

Signed-off-by: Moritz "WanzenBug" Wanzenböck <moritz.wanzenboeck@linbit.com>