chart: always create linstor passphrase
Since LINSTOR can now access S3, it needs a way to securely store these
credentials. This was already supported, using the now confusingely-named
"luksSecret". This secret contains only the passphrase LINSTOR itself uses
to securely store all kinds of secret data, for example LUKS passphrases
(who would have thought...).

Since the actual passphrase itself is not required to be manually configured,
we just create a secret with a random passphrase. We need to be careful with
not re-generating the secret when we upgrade the chart however. Otherwise
we could make some data permanently inaccesible.

Signed-off-by: Moritz "WanzenBug" Wanzenböck <moritz.wanzenboeck@linbit.com>