csidriver: restart node pods if topology labels changed
The satellite operator applies k8s node labels as aux properties. These are
then reported by the csi nodes as supported topology keys. This report
only happens once on plugin startup. Later changes are not reflected, leaving
the volume provisioning steps with old or outdated topology keys.

The only way to force a re-sync of those keys is to delete pod. This change
compares the LINSTOR Node auxilliary properties with the keys reported by
the csi plugin registration. If keys are present in LINSTOR but not
in the registration, the registration and pod is deleted. After the daemonset
restarts the pod, kubelet will trigger a new registration, making the updated
keys available.

Signed-off-by: Moritz "WanzenBug" Wanzenböck <moritz.wanzenboeck@linbit.com>