Automatically create backup of LINSTOR resources
LINSTOR upgrades using the K8s backend are still experimental. While they
should generally wrk, it's better to be safe than sorry. Before an upgrade
the Operator will collect all LINSTOR CRDs and CRs, serialize them as yaml,
and then tar+gzip them so they fit into a secret.

The maximum size of a secret is around 3MB, so there might be clusters
that have more resources than we can fit into the secret. For such
situations the backup is written to the container FS, and the operator will
wait for the user to give the go-ahead by creating the FS.

The backup generation can also be disabled on the operator level.

Signed-off-by: Moritz "WanzenBug" Wanzenböck <moritz.wanzenboeck@linbit.com>