drbd-injection: make source directory configurable
Some distributions don't use /usr/src to store kernel sources. Notably,
Flatcar Container Linux does not have that directory (and it can't easily be
created, either).

This commit adds a new value to the satellite CRD:

  kernelModuleInjectionAdditionalSourceDirectory

If set, the injection container will be started with the directory bound
from the host (read-only). If not set, /usr/src will be bound to keep
existing behaviour. If the given value is not an absolute path, no directory
from the host will be bound (except for /lib/modules, which is required in
all cases).

As a small improvement, we also don't try to mount this directory if
DRBD should not be compiled.

Signed-off-by: Moritz "WanzenBug" Wanzenböck <moritz.wanzenboeck@linbit.com>