csi: make default worker setting backwards compatible
To avoid unnecessary CRD upgrades, the operator uses the zero value
of the worker settings to indicate "use default". This lets the
operator run even when the CRDs were not updated.

Also works around strict helm validation, by only writing the
worker thread setting to CR if it different from the default.

Signed-off-by: Moritz "WanzenBug" Wanzenböck <moritz.wanzenboeck@linbit.com>