fix example deployment generation
helm template seems to separate hooks from other resources, which caused
LinstorController to no longer be present in the examples after moving
the generated secret to a hook. This commit splits the secret from the
controller resource, ensuring both are rendered in the template.

Also adds a warning about deleting or changing the secret.

Signed-off-by: Moritz "WanzenBug" Wanzenböck <moritz.wanzenboeck@linbit.com>