satellite: remove useless readiness probe
The readiness probe was implemented as a connection check on the satellites's
comm-port. This turned out to cause issues, as the satellite only expects a
single connection (from the controller). The most visible issues occured
when using SSL, as the readiness probe would not complete the TLS handshake,
generating repeated error reports.

Since the open comm-port doesn't actually tell us all that much about
the satellites readiness, other than that it completed start-up, we don't
lose very much by just removing it.

Note that a solution querying the Controller API was considered, but discarded,
as it would almost certainly propagate any Controller errors to the
readiness state, causing satellites to be recreated when they actually work
fine.

Signed-off-by: Moritz "WanzenBug" Wanzenböck <moritz.wanzenboeck@linbit.com>