helm: prevent naive passphrase deletion
A user might try to redeploy Piraeus, by using "helm uninstall" and deleting
the whole namespace. This will delete the passphrase secret, but if the user
then tries to redeploy, LINSTOR will not start. That is because the generated
passphrase will be different on redeploy, and LINSTOR compares the passphrase
with whatever is stored in its database. The database is not necessarily
deleted (for example: K8s backend requires manual deletion).

To prevent a user from accidentally deleting the passphrase, we add a finalizer
that is only removed by manual intervention by the user.

Signed-off-by: Moritz "WanzenBug" Wanzenböck <moritz.wanzenboeck@linbit.com>