work around inconsistent linstor flag values
The plain encryption type should have a value of "Plain" according to
linstor-common (used to generate flag values in golinstor), but LINSTOR
actually reports it as "PLAIN". As changing this value in linstor-common
has implications on some LINSTOR internals, changing it was deemed not
worth it. Instead we do a case-insensitive comparison where appropriate.

Signed-off-by: Moritz "WanzenBug" Wanzenböck <moritz.wanzenboeck@linbit.com>