update dependencies
Update all dependencies. Most notabely, this removes code dependency on
operator-sdk. There were also reported crashes fixed in newer versions of
k8s.io/client-go and others.

One change this does introduce is the removal of prometheus operator resources
for the operator itself. These were just the default metrics on the CR, and as
such not that interesting.

Signed-off-by: Moritz "WanzenBug" Wanzenböck <moritz.wanzenboeck@linbit.com>