satellites: only query disks on single satellite
Use new LINSTOR API to only query disks for a single satellite. This reduces
both the load on satellites, and fixes an issue were offline nodes would be
silently ignored.

That offline nodes were ignored led to situations were the storage-pool would
be set up incorrectly. Since no devices was listed while the node was offline,
the operator assumed that the LVM setup was already completed and registered
the storage pool.

The new API in LINSTOR >1.19.0 can be used to query only a single node at a
time, and also reports errors if the requested satellite is offline, preventing
the above error by aborting the storage-pool set-up early.

Signed-off-by: Moritz "WanzenBug" Wanzenböck <moritz.wanzenboeck@linbit.com>