chart: disabled legacy HA Controller deployment
Disabled the legacy HA Controller deployment by default. The new version of
the HA Controller is packaged as a separate chart, to be installed from
the piraeus chart repository.

Also updates the docs accordingly, with instructions on how to use the new
HA Controller, and how to re-enable the legacy version if desired.

Signed-off-by: Moritz "WanzenBug" Wanzenböck <moritz.wanzenboeck@linbit.com>