satellites: only restore evicted sat. on ready nodes
We automatically restore evicted satellites if the pod is running.
Since the status of a pod is responsibility of the kubelet, i.e.
the actual node, an offline node will still show running pods.

In order to prevent restoring satellites that are still offline, we
check the node ready condition in addition to the pod status. This
way we can be reasonably sure that the satellite is actually reachable.

Signed-off-by: Moritz "WanzenBug" Wanzenböck <moritz.wanzenboeck@linbit.com>