Add default image configuration
Adds a subpackages to create kustomize overrides for container images,
based on a simple file-based default configuration.

The configuration sets a specific repository base, and configures the actual
image names + tags for all components in the pkg/resources package (with
the exception of upstream k8s components).

It also supports passing in the ".status.nodeInfo.osImage" information from
the k8s node object. This field seems to be populated by /etc/os-release's
PRETTY_NAME, which we can use to automatically select the right DRBD image
to load.

Signed-off-by: Moritz "WanzenBug" Wanzenböck <moritz.wanzenboeck@linbit.com>