utils: add fieldpath pkg
The fieldpath packages is taken verbatim from k8s.io/pkg/fieldpath,
since it is not available as a standalone packages, and we don't want
to pull in the full kubernetes repository.

It is useful when referencing specific metadata of an object in a "k8s" way,
i.e. it is the syntax used in the so-called "Downward API", which is also useful
for us when referencing specific values a LINSTOR property should inherit from a
k8s node.

Signed-off-by: Moritz "WanzenBug" Wanzenböck <moritz.wanzenboeck@linbit.com>