utils: add kustomize patch wrapper
This commit adds helper for creating and converting patches to a kustomize
compatible format.

We need to support patches originating from our own code, ideally in a
type-safe, and therefore typo-safe way, as well as user patches.

Since we can't use the kustomize types directly in our API (they are missing
some DeepCopy() interfaces), we have to create our own copy for the user-facing
patches. This also gives us the opportunity to get rid of the "path:" patch
type, as the operator can't easily reference files.

For or own patches, we create some wrapper functions that can encode arbitrary
go structs as our patch payload. This makes it possible to use the normal
k8s API definitions as strategic merge patches.

Signed-off-by: Moritz "WanzenBug" Wanzenböck <moritz.wanzenboeck@linbit.com>