utils: add prune helper
Kustomize does not have any built-in way to prune resources that are no longer
needed. We add a helper that accomplishes pruning by listing all potential
resources that need to be pruned, and checking if they should be kept.

We use the OwnerReference on the object to determine if the operator actually
manages the resource. To check for deletion, we check the complete list of
resources that were applied in the last run. Only if a resource is owned
by our operand, but we have not applied it in the last run, the resource will
be pruned.

Signed-off-by: Moritz "WanzenBug" Wanzenböck <moritz.wanzenboeck@linbit.com>