utils: add property helpers
Add API types and helper for dealing with LINSTOR properties. LINSTOR
properties can be defined by the user, but there are also some default
values that the operator will set.

Since the user might want to reference existing values on a k8s node for
satellite and storage pool properties, we add some logic based on the
vendored fieldpath pkg to support this.

Signed-off-by: Moritz "WanzenBug" Wanzenböck <moritz.wanzenboeck@linbit.com>