Add podpatcher utility
In the case of satellites, the operator will have to deal with raw Pods instead
of Deployments or Daemonsets. This is because we have some very particular
requirements:

* We potentially have different Pod Specs per node, for example to use
  different DRBD loader images or include different Satellite configs.
* Satellites should have stable hostnames. This is because LINSTOR does
  not take to kindly to changed "uname"s for deployed resources. Since
  we no longer run in the host network by default, we inherit the Pod name
  as "kernel" hostname.

The issue with using raw Pods is that you can't just patch them like other
resources: Pods are mostly immutable. This could be dealt with quite easily,
if not for the fact that updates to images are allowed.

Updated images would mean that the user would see an unexpected container
restart. In addition, the operator could not easily tell which version
of the container was actually running, as that is not something we can easily
get out of the Pod status.

So our solution is a custom patch function for Pods, that:

* Checks if the Pod already exists
  * If not, just apply the normal patch
* If the new Pod Spec uses the same images, try a normal patch
* If the new Pod Spec uses different images, or the patch could not be
  applied: delete and recreate the Pod.

Signed-off-by: Moritz "WanzenBug" Wanzenböck <moritz.wanzenboeck@linbit.com>