add conditions helper
Conditions are a common way to communicate the state of a k8s resource. We want
to use them for our operator. A usability issue that came up during development
is that the best place to observe a specific condition often varies wildly
based on the exact thing to observe.

The solution is to use our own Conditions helper, which keeps track of a set of
default conditions that all our resources should implement. The Conditions
can be passed around, keeping track of which condition was already observed,
with the intention of letting the operator set the actual conditions at
the end of reconciliation.

Signed-off-by: Moritz "WanzenBug" Wanzenböck <moritz.wanzenboeck@linbit.com>