api: add LinstorSatellite resource
LinstorSatellite defines the desired state of a single LINSTOR satellite in
a cluster. K8s node and Satellite are linked by name, so every LinstorSatellite
resource needs a matching node object.

The resource itself allows setting up storage pools, LINSTOR properties
as well as customizing the default resources by using kustomize patches.

In addition to the resource, we also add a webhook to validate incoming
LinstorSatellite resource. In the webhook we check for common misconfigurations,
so that we can reject them immediately.

Signed-off-by: Moritz "WanzenBug" Wanzenböck <moritz.wanzenboeck@linbit.com>