api: add LinstorSatelliteConfiguration resources
A LinstorSatelliteConfiguration is configuration for LinstorSatellites, that
is applied to one or more (or even all) LinstorSatellites.

The to-be-added LinstorCluster controller will, for each node:
* Merge all LinstorSatelliteConfiguration resources with selectors matching
  the node.
* Create a LinstorSatellite object with the result of that merging process.

As such, the LinstorSatelliteConfiguration has almost the same structure
as a LinstorSatellite, with the addition of the a NodeSelector field. It also
has basically the same additional validation steps performed by the webhook.

Signed-off-by: Moritz "WanzenBug" Wanzenböck <moritz.wanzenboeck@linbit.com>