satellite-configuration: add merge helper
LinstorSatelliteConfigurations apply to one or more satellites. We need to
define a merge strategy, so that a set of Configurations result in a
deterministic and sensible LinstorSatellite resource.

This commit adds a helper method that merges a set of LSConfigurations into
a single LSConfiguration. How the merge happens depends on the field:
* Patches are concatinated, in the original order of configs. The assumption
  being that later patches override earlier patches if required.
* StoragePools are merged by name. If a "later" config has a storge pool
  that shares the name with a storage pool specified earlier, the later
  configuration overrides the first.
* Properties are merged by name. Like with StoragePools, later configurations
  override earlier versions.

This means that the order of LSConfigurations supplied to the helper function
is import. It is the responsibility of the caller to supply the configurations
in a consistent and sensible order.

Signed-off-by: Moritz "WanzenBug" Wanzenböck <moritz.wanzenboeck@linbit.com>