kustomize: mark CRDs as cluster scoped
Our pruning logic compares existing resources to the ones we applied. This
includes the proper namespaces for all resources. By default, kustomize will
apply this namespace to any resource that is not cluster scoped.

The list of cluster scoped resources is baked into kustomize, meaning it also
applied the namespace to LinstorSatellites. This was ignored by the k8s client,
but became important during resource pruning: The pruning helper compared the
deployed (non-namespaced) LinstorSatellite against the kustomize output with
the "fake" namespaced LinstorSatellite, and decided to remove the
LinstorSatellite resource right after creation.

The fix is to add to the known openapi spec used by kustomize to determine
if a resource is namespaced. Now, the kustomize process does not set a
namespace on LinstorSatellites.

Signed-off-by: Moritz "WanzenBug" Wanzenböck <moritz.wanzenboeck@linbit.com>