k8sgc: add utility to emulate k8s garbage collection
In our controller tests, we would like to test deletion of resources
where appropriate. Sadly, the k8s garbage collector is not part of the
API Server, but managed by another component. This means normal
garbage collection via OwnerReferences is not available in the default
test environment.

Since the actual k8s implementation is not available in a reasonably
integrateable fashion, we implement our own basic garbage collector.

Signed-off-by: Moritz "WanzenBug" Wanzenböck <moritz.wanzenboeck@linbit.com>