controllers: set custom rate limit for tests
The default rate limiter for controllers goes up to 1000 seconds back off after
reconcile errors. While this is fine for the actual operators, we don't want
our tests to run into such a delay.

To improve the situation, we replace the controller builder for tests with one
using a custom rate limiter, configured to have a maximum delay of 1 second.

Signed-off-by: Moritz "WanzenBug" Wanzenböck <moritz.wanzenboeck@linbit.com>