olm: ensure generated CSV passes scorecard tests
Ensures that the generated ClusterServiceVersion passes the basic "scorecard"
tests. In particular, we need to ensure we have a basic description of the
status field, as well as a description of all resources generated by a CR.

For the second task, we create a small program that collects that information
directly from the resources embedded in the operator.

Signed-off-by: Moritz "WanzenBug" Wanzenböck <moritz.wanzenboeck@linbit.com>