cluster: add linstor passphrase to API
Add the option to reference a secret holding the LINSTOR master passphrase.
The passphrase is then injected into the LINSTOR Controller pod via environment
variables.

This is quite similar to how it worked in Operator v1. The main difference is
that for v2, we do not plan to automatically create such a secret. This was
always poorly thought out, especially as temporary removal of the deployment
could lead to complete loss of the passphrase. By requiring the user to create
one on their own, they hopefully keep better track of it (or use some external
solution for storing it).

Signed-off-by: Moritz "WanzenBug" Wanzenböck <moritz.wanzenboeck@linbit.com>