patches: convert to yaml-based patches
Having patches to k8s resources directly in the controller code made the code
quite hard to follow.

To replace those patches, use predefined patches, already in the kustomize
format. Since the patches only need setting of single values, we create a
basic "yaml-safe" template system, where we can replace a specific part of
the patch with minimal fuzz.

This commit also converts almost all of the patches to this new system. Only
the LinstorSatellite patch created by the LinstorCluster controller is left
as the old set of manual JSONPatch operations, as it does not fit nicely
in the "replace single parts of patches" framework.

Signed-off-by: Moritz "WanzenBug" Wanzenböck <moritz.wanzenboeck@linbit.com>