v2: reintroduce a helm chart
Reintroduce a Helm Chart, for users that really need it. We don't want to
put too much effot into it, so most of it was created by running
https://github.com/arttor/helmify once.

We have a separate README for the Chart, but otherwise avoid mentioning Helm
in our main guides: new users should be using kustomize.

Signed-off-by: Moritz "WanzenBug" Wanzenböck <moritz.wanzenboeck@linbit.com>