satellite: add drbd-shutdown-guard init-container
The purpose of this container is to install a systemd service unit that runs
during normal shutdown. When the unit is stopped, it will force all resources
into secondary role. This will cause any suspended resource to get "unstuck"
for the purpose of unmounting.

The service is congigured to stop after kubelet, so we can assume that all
still mounted DRBD resources will remain so until systemd does it's normal
clean up. It also means that the network for DRBD is likely already torn
down, so any pending writes would be lost in any case.

Signed-off-by: Moritz "WanzenBug" Wanzenböck <moritz.wanzenboeck@linbit.com>