how-to: run on MicroK8s
MicroK8s needs slightly modified mount paths for the CSI node containers.

During testing it also revealed an issue in the way we pass the csi socket to
the linstor-csi plugin: instead of using the explicit mount location it used
a path for the csi socket that only worked by accident. To fix it, we now
use the explicit socket mount location.

Signed-off-by: Moritz "WanzenBug" Wanzenböck <moritz.wanzenboeck@linbit.com>