operator: run reconiliation on changed image config
The image configuration can be changed without the Operator restarting. But
the Operator would never reload the config, so the changes would only take
effect after restarting the Pod.

This commit adds support for reading the image configuration directly from
the kubernetes API, also triggering a reload automatically when a changed
config is detected.

Signed-off-by: Moritz "WanzenBug" Wanzenböck <moritz.wanzenboeck@linbit.com>