chart: remove missing config
In 5f77f51 we reworked the way the image config is applied to the operator.
We forgot to remove the directly mounted config map from the chart deployment.

This commit removes the config map mount, as it no longer required: the
operator will use the k8s API directly to fetch the image config.

Signed-off-by: Moritz "WanzenBug" Wanzenböck <moritz.wanzenboeck@linbit.com>